move languages to dna, move readConfig out of languages, read language from config
