+++ src/languages/nodejs.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const util = require('util')
 const path = require('path')
-const { confirmWriteFile } = require('../util')
+const { confirmWriteFile, readConfig } = require('../util')
 
 const readFile = util.promisify(fs.readFile)
 
@@ -30,18 +30,8 @@ COPY --chown=nodejs . ./
 CMD ["node", "${entrypoint}"]`
   },
 
-  readConfig: async (options) => {
-    let packageJson = {}
-    try {
-      packageJson = JSON.parse(await readFile(path.join(options.directory, './package.json')))
-    } catch (_err) {}
-    const config = packageJson['deploy-node-app'] || {}
-    if (!config.name) config.name = packageJson.name
-    return config
-  },
-
   writeConfig: async function (config, options) {
-    const packageJson = await this.readConfig()
+    const packageJson = await readConfig()
     packageJson['deploy-node-app'] = config
     await confirmWriteFile('./package.json', JSON.stringify(packageJson, null, 2) + '\n', {
       ...options,
