+++ src/languages/php.js
@@ -1,4 +1,3 @@
-// const fs = require('fs')
 
 module.exports = {
   name: 'php',
